+++ utils-project/gradle-launcher/src/main/java/com/asakusafw/utils/gradle/CommandLauncher.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright 2011-2017 Asakusa Framework Team.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.asakusafw.utils.gradle;
+
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Launches commands.
+ * @since 0.9.2
+ */
+@FunctionalInterface
+public interface CommandLauncher {
+
+    /**
+     * Launches the command.
+     * @param command the command location
+     * @param arguments the command arguments
+     * @return the exit status
+     */
+    int launch(Path command, List<String> arguments);
+
+    /**
+     * Launches the command.
+     * @param command the command location
+     * @param arguments the command arguments
+     * @return the exit status
+     */
+    default int launch(Path command, String... arguments) {
+        return launch(command, Arrays.asList(arguments));
+    }
+}
