+++ utils-project/gradle-launcher/src/test/data/base-package/bin/make_output.sh
